microservices/ai_assistant/requirements.txt
Flask==2.3.3
pymongo==4.6.1
dnspython==2.4.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
stripe==7.8.0
eventlet==0.33.3
gunicorn==21.2.0